--- src/main/java/Command.java
@@ -1,7 +1,0 @@
-public abstract class Command {
-
-    abstract public void execute(Storage storage, TaskList tasks, Ui ui) throws OptimusExceptions;
-    public boolean shouldContinue() {
-       return true;
-    }
-}
