--- src/main/java/IncompleteCommandException.java
@@ -1,5 +1,0 @@
-public class IncompleteCommandException extends OptimusExceptions{
-    public IncompleteCommandException(String message) {
-        super(message);
-    }
-}
