--- src/main/java/InvalidCommandException.java
@@ -1,6 +1,0 @@
-public class InvalidCommandException extends OptimusExceptions{
-
-    public InvalidCommandException(String message) {
-        super(message);
-    }
-}
