--- src/main/java/InvalidDateFormatException.java
@@ -1,5 +1,0 @@
-public class InvalidDateFormatException extends OptimusExceptions{
-    public InvalidDateFormatException(String message) {
-        super(message);
-    }
-}
