--- src/main/java/InvalidTaskNumberException.java
@@ -1,9 +1,0 @@
-public class InvalidTaskNumberException extends OptimusExceptions{
-    public InvalidTaskNumberException() {
-        super("A Task with this number does not exist.");
-    }
-
-    public InvalidTaskNumberException(String message) {
-        super(message);
-    }
-}
