+++ src/main/java/optimus/commands/ListCommand.java
@@ -1,4 +1,12 @@
-public class ListCommand  extends Command{
+package optimus.commands;
+
+import optimus.commands.Command;
+import optimus.Storage;
+import optimus.TaskList;
+import optimus.Ui;
+import optimus.exceptions.InvalidTaskNumberException;
+
+public class ListCommand  extends Command {
 
     public ListCommand() {
