+++ src/main/java/optimus/commands/UnmarkCommand.java
@@ -1,4 +1,13 @@
-public class UnmarkCommand extends Command{
+package optimus.commands;
+
+import optimus.commands.Command;
+import optimus.Storage;
+import optimus.TaskList;
+import optimus.Ui;
+import optimus.exceptions.InvalidTaskNumberException;
+import optimus.tasks.Task;
+
+public class UnmarkCommand extends Command {
     private final int taskNum;
 
     public UnmarkCommand(int taskNum) {
