+++ src/main/java/optimus/exceptions/IncompleteCommandException.java
@@ -1,0 +1,9 @@
+package optimus.exceptions;
+
+import optimus.exceptions.OptimusExceptions;
+
+public class IncompleteCommandException extends OptimusExceptions {
+    public IncompleteCommandException(String message) {
+        super(message);
+    }
+}
