+++ src/main/java/optimus/exceptions/InvalidCommandException.java
@@ -1,0 +1,10 @@
+package optimus.exceptions;
+
+import optimus.exceptions.OptimusExceptions;
+
+public class InvalidCommandException extends OptimusExceptions {
+
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
