+++ src/main/java/optimus/exceptions/InvalidDateFormatException.java
@@ -1,0 +1,9 @@
+package optimus.exceptions;
+
+import optimus.exceptions.OptimusExceptions;
+
+public class InvalidDateFormatException extends OptimusExceptions {
+    public InvalidDateFormatException(String message) {
+        super(message);
+    }
+}
