+++ src/main/java/optimus/exceptions/InvalidTaskNumberException.java
@@ -1,0 +1,13 @@
+package optimus.exceptions;
+
+import optimus.exceptions.OptimusExceptions;
+
+public class InvalidTaskNumberException extends OptimusExceptions {
+    public InvalidTaskNumberException() {
+        super("A optimus.tasks.Task with this number does not exist.");
+    }
+
+    public InvalidTaskNumberException(String message) {
+        super(message);
+    }
+}
