+++ src/main/java/optimus/exceptions/OptimusExceptions.java
@@ -1,3 +1,5 @@
+package optimus.exceptions;
+
 public class OptimusExceptions extends Exception{
     public OptimusExceptions(String message) {
         super(message);
