+++ src/main/java/optimus/tasks/DeadlineTask.java
@@ -1,12 +1,21 @@
+package optimus.tasks;
+
+import optimus.Storage;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.Objects;
 
-public class ToDoTask extends Task {
-    public ToDoTask(String desc) {
+public class DeadlineTask extends Task {
+
+    LocalDate deadline;
+    public DeadlineTask(String desc, LocalDate deadline) {
         super(desc);
+        this.deadline = deadline;
     }
 
-    public ToDoTask(String desc, String status) {
-        this(desc);
+    public DeadlineTask(String desc, LocalDate deadline, String status) {
+        this(desc, deadline);
         if (Objects.equals(status, "1")) {
             super.markAsComplete();
         }
@@ -15,20 +24,23 @@ public class ToDoTask extends Task {
     @Override
     public String toString() {
         String s = "";
-        s += "[T]";
+        s += "[D]";
         s += "[" + super.getStatusString() + "] ";
         s += super.getTaskDesc();
+        s += String.format(" (by: %s)", this.deadline.format(DateTimeFormatter.ofPattern("MMM d yyyy")));
         return s;
     }
 
     @Override
     public String getStorageString() {
         String s = "";
-        s += "T";
+        s += "D";
         s += Storage.SPECIAL_CHAR;
         s += super.getStatusInt();
         s += Storage.SPECIAL_CHAR;
         s += super.getTaskDesc();
+        s += Storage.SPECIAL_CHAR;
+        s += deadline;
         s += Storage.SPECIAL_CHAR;
         return s;
     }
