+++ src/main/java/optimus/tasks/EventTask.java
@@ -1,3 +1,7 @@
+package optimus.tasks;
+
+import optimus.Storage;
+
 import java.util.Objects;
 
 public class EventTask extends Task {
