+++ src/main/java/optimus/tasks/ToDoTask.java
@@ -1,17 +1,17 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
+package optimus.tasks;
+
+import optimus.Storage;
+import optimus.tasks.Task;
+
 import java.util.Objects;
 
-public class DeadlineTask extends Task{
-
-    LocalDate deadline;
-    public DeadlineTask(String desc, LocalDate deadline) {
+public class ToDoTask extends Task {
+    public ToDoTask(String desc) {
         super(desc);
-        this.deadline = deadline;
     }
 
-    public DeadlineTask(String desc, LocalDate deadline, String status) {
-        this(desc, deadline);
+    public ToDoTask(String desc, String status) {
+        this(desc);
         if (Objects.equals(status, "1")) {
             super.markAsComplete();
         }
@@ -20,23 +20,20 @@ public class DeadlineTask extends Task{
     @Override
     public String toString() {
         String s = "";
-        s += "[D]";
+        s += "[T]";
         s += "[" + super.getStatusString() + "] ";
         s += super.getTaskDesc();
-        s += String.format(" (by: %s)", this.deadline.format(DateTimeFormatter.ofPattern("MMM d yyyy")));
         return s;
     }
 
     @Override
     public String getStorageString() {
         String s = "";
-        s += "D";
+        s += "T";
         s += Storage.SPECIAL_CHAR;
         s += super.getStatusInt();
         s += Storage.SPECIAL_CHAR;
         s += super.getTaskDesc();
-        s += Storage.SPECIAL_CHAR;
-        s += deadline;
         s += Storage.SPECIAL_CHAR;
         return s;
     }
